backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0
PyJWT==2.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.1
marshmallow==3.20.1
gmqtt==0.6.16
websockets==12.0
flasgger==0.9.7.1
